+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/Webhook/message_handlers.yml
@@ -1,0 +1,6 @@
+services:
+    Akeneo\Connectivity\Connection\Infrastructure\Webhook\MessageHandler\BusinessEventHandler:
+        arguments:
+            - '@Akeneo\Connectivity\Connection\Application\Webhook\Command\SendBusinessEventToWebhooksHandler'
+            - '@event_dispatcher'
+        tags: [messenger.message_handler]
