+++ src/Akeneo/Connectivity/Connection/back/tests/EndToEnd/Webhook/ConsumeProductModelEventEndToEnd.php
@@ -6,7 +6,7 @@ namespace Akeneo\Connectivity\Connection\back\tests\EndToEnd\Webhook;
 
 use Akeneo\Connectivity\Connection\Domain\Settings\Model\Read\ConnectionWithCredentials;
 use Akeneo\Connectivity\Connection\Domain\Settings\Model\ValueObject\FlowType;
-use Akeneo\Connectivity\Connection\Infrastructure\MessageHandler\BusinessEventHandler;
+use Akeneo\Connectivity\Connection\Infrastructure\Webhook\MessageHandler\BusinessEventHandler;
 use Akeneo\Pim\Enrichment\Component\Product\Message\ProductModelCreated;
 use Akeneo\Pim\Enrichment\Component\Product\Message\ProductModelRemoved;
 use Akeneo\Pim\Enrichment\Component\Product\Message\ProductModelUpdated;
@@ -57,7 +57,7 @@ class ConsumeProductModelEventEndToEnd extends ApiTestCase
             ]
         );
 
-        /** @var $businessEventHandler BusinessEventHandler */
+        /** @var $businessEventHandler \Akeneo\Connectivity\Connection\Infrastructure\Webhook\MessageHandler\BusinessEventHandler */
         $businessEventHandler = $this->get(BusinessEventHandler::class);
         $businessEventHandler->__invoke($message);
 
@@ -90,7 +90,7 @@ class ConsumeProductModelEventEndToEnd extends ApiTestCase
             ]
         );
 
-        /** @var $businessEventHandler BusinessEventHandler */
+        /** @var $businessEventHandler \Akeneo\Connectivity\Connection\Infrastructure\Webhook\MessageHandler\BusinessEventHandler */
         $businessEventHandler = $this->get(BusinessEventHandler::class);
         $businessEventHandler->__invoke($message);
