+++ src/Akeneo/Connectivity/Connection/back/tests/Unit/spec/Infrastructure/Webhook/MessageHandler/BusinessEventHandlerSpec.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
-namespace spec\Akeneo\Connectivity\Connection\Infrastructure\MessageHandler;
+namespace spec\Akeneo\Connectivity\Connection\Infrastructure\Webhook\MessageHandler;
 
 use Akeneo\Connectivity\Connection\Application\Webhook\Command\SendBusinessEventToWebhooksCommand;
 use Akeneo\Connectivity\Connection\Application\Webhook\Command\SendBusinessEventToWebhooksHandler;
 use Akeneo\Connectivity\Connection\Domain\Webhook\Event\MessageProcessedEvent;
+use Akeneo\Connectivity\Connection\Infrastructure\Webhook\MessageHandler\BusinessEventHandler;
 use Akeneo\Platform\Component\EventQueue\BulkEvent;
 use Akeneo\Platform\Component\EventQueue\BulkEventInterface;
 use PhpSpec\ObjectBehavior;
@@ -30,7 +31,7 @@ class BusinessEventHandlerSpec extends ObjectBehavior
 
     public function it_is_initializable(): void
     {
-        $this->shouldHaveType(\Akeneo\Connectivity\Connection\Infrastructure\Webhook\MessageHandler\BusinessEventHandler::class);
+        $this->shouldHaveType(BusinessEventHandler::class);
         $this->shouldImplement(MessageSubscriberInterface::class);
     }
